backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
bcrypt
python-jose[cryptography]
python-multipart
